requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.11.12
environs==14.5.0
frozenlist==1.8.0
idna==3.11
magic-filter==1.0.12
marshmallow==4.1.0
multidict==6.7.0
propcache==0.4.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.2.1
typing_extensions==4.15.0
yarl==1.22.0
